Move isWord to lib/is-word module
